moved schema classes to common
