+++ common/client/src/test/java/zingg/common/client/schema/SchemaClusterNull.java
@@ -1,0 +1,15 @@
+package zingg.common.client.schema;
+
+public class SchemaClusterNull {
+    public final Integer z_z_zid;
+    public final Integer z_cluster;
+    public final Integer z_score;
+    public final String z_zsource;
+
+    public SchemaClusterNull(Integer z_z_zid, Integer z_cluster, Integer z_score, String z_zsource) {
+        this.z_z_zid = z_z_zid;
+        this.z_cluster = z_cluster;
+        this.z_score = z_score;
+        this.z_zsource = z_zsource;
+    }
+}
